+++ Stream-O-Vision/SqliteHandler.h
@@ -17,13 +17,13 @@ class SqliteHandler
 
 		BOOL LoadAll();
 		std::vector<Station> GetStationMap();
-		BOOL GetPlaylist(int,int);
+		BOOL GetPlaylist(int64_t,int);
 
-		int AddStation(std::string, std::string);
-		BOOL DeleteStation(int);
-		BOOL DeletePlaylistByStationId(int);
-		BOOL DeletePlaylistItemById(int);
-		int AddPlaylistItem(std::string, std::string, int);
+		int64_t AddStation(std::string, std::string);
+		BOOL DeleteStation(int64_t);
+		BOOL DeletePlaylistByStationId(int64_t);
+		BOOL DeletePlaylistItemById(int64_t);
+		int64_t AddPlaylistItem(std::string, std::string, int64_t);
 
 		BOOL CreateTables();
